requirements.txt
PyGithub
pytest